fixed a bug with larger data

+++ Shadow/src/main/java/org/operatorfoundation/shapeshifter/shadow/kotlin/ShadowInputStream.kt
@@ -64,8 +64,8 @@ class ShadowInputStream(private val networkInputStream: InputStream, private val
         val lengthData = decryptionCipher.decrypt(encryptedLengthData)
 
         // change lengthData from BigEndian representation to int length
-        val leftByte = lengthData[0]
-        val rightByte = lengthData[1]
+        val leftByte = lengthData[0].toUByte()
+        val rightByte = lengthData[1].toUByte()
         val rightInt = rightByte.toInt()
         val leftInt = leftByte.toInt()
         val payloadLength = (leftInt * 256) + rightInt
